appsflyer/requirements.txt
# AppsFlyer ingestion (aggregate Pull API — CSV over HTTP)
python-dotenv>=1.0.0
requests>=2.31.0
